refactoring cli
